refactor(contentful): move csv-import classes to different files
