+++ packages/botonic-plugin-contentful/src/tools/l10n/import-csv-from-translators.ts
@@ -3,11 +3,9 @@ import { ManageContentful } from '../../contentful/manage'
 import { ManageContext } from '../../manage-cms'
 import { ContentfulOptions } from '../../plugin'
 import { isOfType } from '../../util/enums'
-import {
-  CsvImport,
-  ReferenceFieldDuplicator,
-  StringFieldImporter,
-} from './csv-import'
+import { CsvImport } from './csv-import'
+import { ReferenceFieldDuplicator } from './reference-field-duplicator'
+import { StringFieldImporter } from './string-field-importer'
 
 async function readCsvForTranslators(
   contentfulOptions: ContentfulOptions,
